requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
openai==1.55.0
python-dotenv==1.0.1
pdf2image==1.17.0